+++ auto_factory/lib/src/entry_point.dart
@@ -1,14 +1,14 @@
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
-import './provider_generator.dart';
+import './generator.dart';
 
 /// Provider called by build runner
 Builder autoFactoryProviderBuilder(BuilderOptions options) => SharedPartBuilder(
       [
         ProviderGenerator(),
-        FactoryGenerator(),
-        CachingFactoryGenerator(),
+        InjectableGenerator(),
+        SingletonFactoryGenerator(),
       ],
       'auto_factory',
     );