+++ auto_factory/test/auto_factory_test.dart
@@ -1,16 +1,16 @@
-import 'package:auto_factory/src/provider_generator.dart';
+import 'package:auto_factory/src/generator.dart';
 import 'package:path/path.dart' as path;
 import 'package:source_gen_test/source_gen_test.dart';
 
 Future<void> main() async {
   initializeBuildLogTracking();
-
+/*
   testAnnotatedElements(
     await initializeLibraryReaderForDirectory(
       path.join('test', 'sources'),
       '_basic.dart',
     ),
-    FactoryGenerator(),
+    InjectableGenerator(),
   );
 
   testAnnotatedElements(
@@ -18,6 +18,14 @@ Future<void> main() async {
       path.join('test', 'sources'),
       '_caching.dart',
     ),
-    CachingFactoryGenerator(),
+    SingletonFactoryGenerator(),
+  );
+*/
+  testAnnotatedElements(
+    await initializeLibraryReaderForDirectory(
+      path.join('test', 'sources'),
+      '_sync.dart',
+    ),
+    InjectableGenerator(),
   );
 }