+++ auto_factory/test/sources/_basic_dep.dart
@@ -14,7 +14,7 @@ class DependencyFactory {
   }
 }
 ''')
-@Injectable()
+@Injectable(asyncMode: true)
 class Dependency {
   SecondDependency secondDependency;