+++ auto_factory/test/sources/_basic_second_dep.dart
@@ -8,5 +8,5 @@ class SecondDependencyFactory {
   }
 }
 ''')
-@Injectable()
+@Injectable(asyncMode: true)
 class SecondDependency {}