+++ auto_factory/test/sources/_sync.dart
@@ -1,27 +1,26 @@
 import 'package:auto_factory_annotation/auto_factory_annotation.dart';
 import 'package:source_gen_test/annotations.dart';
 
-import '_basic_dep.dart';
-import '_basic_provider.dart';
-import '_basic_third_dep.dart';
+import '_sync_dep.dart';
+import '_sync_provider.dart';
 
 @ShouldGenerate('''
 class MyComponentFactory {
-  Future<MyComponent> create() async {
+  MyComponent create() {
     final dependencyFactory = DependencyFactory();
     final thirdDependencyProviderFactory = ThirdDependencyProviderFactory();
 
     return MyComponent(
-      await dependencyFactory.create(),
-      await (await thirdDependencyProviderFactory.create()).provide(),
+      dependencyFactory.create(),
+      thirdDependencyProviderFactory.create().provide(),
     );
   }
 }
 ''')
-@Injectable(asyncMode: true)
+@Injectable(asyncMode: false)
 class MyComponent {
   final Dependency dep;
-  final ThirdDependency third;
+  final String third;
 
   MyComponent(
     this.dep,