+++ auto_factory/test/sources/_sync_dep.dart
@@ -1,0 +1,12 @@
+import 'package:auto_factory_annotation/auto_factory_annotation.dart';
+import 'package:source_gen_test/annotations.dart';
+
+@ShouldGenerate('''
+class DependencyFactory {
+  Dependency create()  {
+    return Dependency();
+  }
+}
+''')
+@Injectable(asyncMode: false)
+class Dependency {}