+++ auto_factory/test/sources/_sync_provider.dart
@@ -1,0 +1,9 @@
+import 'package:auto_factory_annotation/auto_factory_annotation.dart';
+
+@Provider(asyncMode: false)
+class ThirdDependencyProvider implements ProviderBase<String> {
+  @override
+  String provide() {
+    return 'aString';
+  }
+}