+++ auto_factory_annotation/example/lib/src/provider/name_provider.g.dart
@@ -1,0 +1,25 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'name_provider.dart';
+
+// **************************************************************************
+// ProviderGenerator
+// **************************************************************************
+
+class NameProviderFactory {
+  static final NameProviderFactory _singleton = NameProviderFactory._internal();
+
+  factory NameProviderFactory() {
+    return _singleton;
+  }
+
+  NameProviderFactory._internal();
+
+  NameProvider? _objectInstance;
+
+  Future<NameProvider> create() async {
+    _objectInstance ??= NameProvider();
+
+    return _objectInstance!;
+  }
+}