+++ auto_factory_annotation/example/main.dart
@@ -1,0 +1,7 @@
+import 'lib/src/application.dart';
+
+void main() async {
+  final app = await ApplicationFactory().create();
+
+  app.greet();
+}