+++ auto_factory_annotation/lib/auto_factory_annotation.dart
@@ -2,17 +2,20 @@ import 'dart:async';
 
 /// Builds a factory shared part and so makes class injectable
 class Injectable {
-  const Injectable();
+  final bool asyncMode;
+  const Injectable({this.asyncMode = false});
 }
 
 /// Same as Component(), but the class instance is cached
 class Singleton {
-  const Singleton();
+  final bool asyncMode;
+  const Singleton({this.asyncMode = false});
 }
 
 /// Creates a provier. should implement the ProviderBase<T> abstract class
 class Provider {
-  const Provider();
+  final bool asyncMode;
+  const Provider({this.asyncMode = false});
 }
 
 abstract class ProviderBase<T> {