+++ auto_factory_annotation/CHANGELOG.md
@@ -22,3 +22,12 @@
 ## 2.0.0
 
 - null safety
+
+## 2.0.0+1
+
+- Added example
+
+
+## 2.1.0 2.1.1
+
+- Added sync / async choiche